+++ lib/commandline/cadd.ml
@@ -71,8 +71,9 @@ let man =
       "If one of the following option is provided $(b,-d, -e, -l, -s, -u), \
        cithare-add(1) will automatically generate a password and the options \
        to generate the password are the same than cithare-generate-password(1)";
-    `P 
-      "$(b,-c) option is only relevant and used if cithare-add(1) will automatically generate a password";
+    `P
+      "$(b,-c) option is only relevant and used if cithare-add(1) will \
+       automatically generate a password";
     `S Manpage.s_see_also;
     `P "cithare-generate-password(1)";
   ]